Rationalize linkage to XSI libraries
The xsi_loader.cpp / xsi_loader.h provided by Xilinx in their examples
is, in hindsight, not intended to be the actual interface to XSI. It's
just a serving suggestion (and not a terribly good API).

Instead, much better to make the shared library (only!) the API
interface. They provide xsi.h to ensure correct function typing.
Otherwise, we don't want their loader code - for example, it doesn't
raise exceptions when it should, it raises compiler warnings, and it
shuffles around the API in ways that are unnecessary.